Fix header tag in default layout
The header now has two classes, one that is used when the page
is on index.html and another that is used at any point. The
index has a unique header as it will be bigger to highlight the
blog's title and won't require an anchor back to the same page.